project 폴더 변경 및 수정
