--- mincheol/Week01/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'Week01'
