--- mincheol/Week01/src/main/java/LV0/adddigits/AddingDigits.java
@@ -1,8 +1,0 @@
-package LV0.adddigits;
-
-public class AddingDigits {
-    public static void main(String[] args) {
-        System.out.println(Solution.solution(1234));
-        System.out.println(Solution.solution(930211));
-    }
-}
