--- mincheol/Week01/src/main/java/LV0/bacterialgrowth/BacterialGrowth.java
@@ -1,8 +1,0 @@
-package LV0.bacterialgrowth;
-
-public class BacterialGrowth {
-    public static void main(String[] args) {
-        System.out.println(Solution.solution(2, 10));
-        System.out.println(Solution.solution(7, 15));
-    }
-}
