--- mincheol/Week01/src/main/java/LV0/distributepizza/PizzaDistributing.java
@@ -1,9 +1,0 @@
-package LV0.distributepizza;
-
-public class PizzaDistributing {
-    public static void main(String[] args) {
-        System.out.println(Solution.solution(7));
-        System.out.println(Solution.solution(1));
-        System.out.println(Solution.solution(15));
-    }
-}
