--- mincheol/Week01/src/main/java/LV0/letter/Letter.java
@@ -1,8 +1,0 @@
-package LV0.letter;
-
-public class Letter {
-    public static void main(String[] args) {
-        System.out.println(Solution.solution("happy birthday!"));
-        System.out.println(Solution.solution("I love you~"));
-    }
-}
