--- mincheol/Week02/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Aug 09 09:22:40 KST 2025
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.14-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
