--- mincheol/Week02/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'Week02'
