+++ mincheol/hwcamp-coding-test-study/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'hwcamp-coding-test-study'
