+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/adddigits/AddingDigits.java
@@ -1,0 +1,8 @@
+package Week01.LV0.adddigits;
+
+public class AddingDigits {
+    public static void main(String[] args) {
+        System.out.println(Solution.solution(1234));
+        System.out.println(Solution.solution(930211));
+    }
+}
