+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/bacterialgrowth/BacterialGrowth.java
@@ -1,0 +1,8 @@
+package Week01.LV0.bacterialgrowth;
+
+public class BacterialGrowth {
+    public static void main(String[] args) {
+        System.out.println(Solution.solution(2, 10));
+        System.out.println(Solution.solution(7, 15));
+    }
+}
