+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/bacterialgrowth/Solution.java
@@ -1,0 +1,15 @@
+package Week01.LV0.bacterialgrowth;
+
+/*
+ * 이름: 변민철
+ * 날짜: 25. 8. 5.
+ * 문제: https://school.programmers.co.kr/learn/courses/30/lessons/120910
+ * */
+
+public class Solution {
+    public static int solution(int n, int t) {
+        int answer = 0;
+        answer = n * (int)Math.pow(2, t);
+        return answer;
+    }
+}
