+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/distributepizza/PizzaDistributing.java
@@ -1,0 +1,9 @@
+package Week01.LV0.distributepizza;
+
+public class PizzaDistributing {
+    public static void main(String[] args) {
+        System.out.println(Solution.solution(7));
+        System.out.println(Solution.solution(1));
+        System.out.println(Solution.solution(15));
+    }
+}
