+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/letter/Letter.java
@@ -1,0 +1,8 @@
+package Week01.LV0.letter;
+
+public class Letter {
+    public static void main(String[] args) {
+        System.out.println(Solution.solution("happy birthday!"));
+        System.out.println(Solution.solution("I love you~"));
+    }
+}
