+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/letter/Solution.java
@@ -1,0 +1,15 @@
+package Week01.LV0.letter;
+
+/*
+ * 이름: 변민철
+ * 날짜: 25. 8. 5.
+ * 문제: https://school.programmers.co.kr/learn/courses/30/lessons/120898
+ * */
+
+public class Solution {
+    public static int solution(String message) {
+        int answer = 0;
+        answer = message.length() * 2;
+        return answer;
+    }
+}
