+++ mincheol/hwcamp-coding-test-study/src/main/java/Week01/LV0/makemaxvalue/MakingMaxValue.java
@@ -1,0 +1,8 @@
+package Week01.LV0.makemaxvalue;
+
+public class MakingMaxValue {
+    public static void main(String[] args) {
+        System.out.println(Solution.solution(new int[]{1, 2, 3, 4, 5}));
+        System.out.println(Solution.solution(new int[]{0, 31, 24, 10, 1, 9}));
+    }
+}
