+++ mincheol/hwcamp-coding-test-study/src/main/java/Week02/LV1/nonfinishedplayer/Solution.java
@@ -1,0 +1,9 @@
+package LV1.nonfinishedplayer;
+
+public class Solution {
+    public static String solution(String[] participant, String[] completion) {
+        String answer = "";
+
+        return answer;
+    }
+}
