+++ mincheol/hwcamp-coding-test-study/src/main/java/Week02/LV2/fibonaccinumbers/FibonacciNumbers.java
@@ -1,0 +1,4 @@
+package Week02.LV2.fibonaccinumbers;
+
+public class FibonacciNumbers {
+}
